Reshape the identity package API for clarity
