+++ docs/architecture.md
@@ -72,8 +72,8 @@ Architectural decisions and behavioral constraints for Linky. Keep this file up 
 - Unauthenticated restore now opens a dedicated `I'm returning` onboarding step with a password-style autofocused SLIP-39 input, manual/paste entry, inline separator normalization, live word validation, and clipboard-triggered auto-submit when a full valid share is pasted
 - New account creation now pauses before first login on an unauthenticated profile-picker step: it derives one deterministic DiceBear avatar from the freshly generated `npub`, picks the suggested display name from a language-matched Czech or English name list based on the current app language, lets the user edit that name immediately, offers 8 trait-group edit buttons (top, hair color, accessories, eyes/brows, mouth, facial hair, skin, clothing) to refine that one avatar, supports uploading a custom square-cropped photo as a 9th option, and the welcome/unauthenticated screens expose the same language menu as the later onboarding steps; the visible confirm action now request-submits a separate hidden credential form containing the chosen name plus a generated 20-word SLIP-39 seed to maximize browser password-manager save heuristics before publishing initial kind-0 metadata and entering the app; the published metadata still includes name, picture, and the default `${npub}@linky.fit` `lud16`
 - Cashu deterministic wallet seed is derived from the SLIP-39 secret using **BIP-85** at path `m/83696968'/39'/0'/24'/0'` (24-word mnemonic)
-- Web app seed/identity helpers in `src/utils/slip39Nostr.ts` are app-level adapters that delegate SLIP-39/BIP-85 derivation to `@linky/identity/identity`. `@linky/identity` depends on `@linky/linkstr` for the nostr key layer — `IdentityProvider` returns linkstr's `NostrSecretKey`/`Pubkey` brands and nsec/npub encoding uses linkstr's codecs; identity owns only derivation (SLIP-39, BIP-32/85, owner lanes) and has no direct `nostr-tools` dependency
-- Seed derivation is a cross-app compatibility contract with Payky (`finitoapp/payky`): the same SLIP-39 share must recover the same master secret in both apps, and the cashu path `m/83696968'/39'/0'/24'/0'` must yield the same wallet. A shared vector table is pinned in `packages/identity/src/identity/crossAppDerivation.test.ts` here and mirrored in Payky's `src/core/modules/shared/cross-app-derivation.test.ts`; intentional derivation changes must update both copies in the same way
+- Web app seed/identity helpers in `src/utils/slip39Nostr.ts` are app-level adapters that delegate SLIP-39/BIP-85 derivation to `@linky/identity`. `@linky/identity` depends on `@linky/linkstr` for the nostr key layer — `IdentityProvider` returns linkstr's `NostrSecretKey`/`Pubkey` brands and nsec/npub encoding uses linkstr's codecs; identity owns only derivation (SLIP-39, BIP-32/85, owner lanes) and has no direct `nostr-tools` dependency
+- Seed derivation is a cross-app compatibility contract with Payky (`finitoapp/payky`): the same SLIP-39 share must recover the same master secret in both apps, and the cashu path `m/83696968'/39'/0'/24'/0'` must yield the same wallet. A shared vector table is pinned in `packages/identity/src/crossAppDerivation.test.ts` here and mirrored in Payky's `src/core/modules/shared/cross-app-derivation.test.ts`; intentional derivation changes must update both copies in the same way
 - Own profile uses a dedicated `#profile/edit` route for editing; the top-left profile avatar modal remains for viewing QR/profile share info and its edit action now opens that standalone page instead of inline modal editing, and `#profile/claim-lightning` lets users check availability and price for a custom `@linky.fit` lightning address, pay the quote in-app, publish the chosen `lud16`, and re-activate the currently paid hosted alias later with one tap; the profile editor hides the default `Restore` button while such a paid alias exists
 - Hosted npub.cash-compatible claim/info/mint sync resolves the API origin from the active lightning-address domain; current mapping keeps default `@npub.cash` on `https://npub.cash` and routes `@linky.fit` through `https://npub.linky.fit`
 - When a user owns a paid `@linky.fit` alias, hosted mint sync for that alias must keep targeting `https://npub.linky.fit` even if the current profile `lud16` points somewhere else; if the hosted mint update fails, Linky must not persist the new local main-mint choice as if it succeeded
