+++ packages/identity/package.json
@@ -7,8 +7,7 @@
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts",
-    "./identity": "./src/identity/index.ts"
+    ".": "./src/index.ts"
   },
   "scripts": {
     "eslint": "eslint --fix .",
