--- packages/identity/src/identity/index.ts
@@ -1,4 +1,0 @@
-export * from "./domain";
-export * from "./IdentityProvider";
-export * from "./MasterSecretProvider";
-export * from "./utils";
